Bug 1748718 - Part 3: Remove now-unnecessary xpidl constant casts, r=emilio
Differential Revision: https://phabricator.services.mozilla.com/D135166

UltraBlame original commit: 32873022d2ed99f8a4d29df5e60c96f3dc8ff9f3

+++ toolkit/components/glean/api/src/ipc.rs
@@ -82,7 +82,7 @@ static PROCESS_TYPE: Lazy<AtomicU32> = Lazy::new(|| {
     
     let process_type = process_type
         .try_into()
-        .unwrap_or(nsIXULRuntime::PROCESS_TYPE_DEFAULT as u32);
+        .unwrap_or(nsIXULRuntime::PROCESS_TYPE_DEFAULT);
     
     
     
@@ -93,7 +93,7 @@ static PROCESS_TYPE: Lazy<AtomicU32> = Lazy::new(|| {
 
 #[cfg(feature = "with_gecko")]
 pub fn need_ipc() -> bool {
-    PROCESS_TYPE.load(Ordering::Relaxed) != nsIXULRuntime::PROCESS_TYPE_DEFAULT as u32
+    PROCESS_TYPE.load(Ordering::Relaxed) != nsIXULRuntime::PROCESS_TYPE_DEFAULT
 }
